wip: Update to ChipDNA 3.14, move files around, test on production
