+++ fattmerchant-ios-sdk/Jobs/CardReader/DisconnectCardReaderJob.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+class DisconnectCardReaderJob : Job {
+  typealias Output = Bool
+  
+  func start() async throws -> Bool {
+    guard let card = Services.resolve(CardReaderService.self) else {
+      throw StaxInitializeException.missingInitializationDetails
+    }
+
+    return try await card.disconnectFromReader()
+  }
+}
